Changes output of SequenceArray.parse_fasta and SequenceAlignment.parse_fasta
Before SequenceArray.parse_fasta returns an OrderedDict of record names and sequences.
This commit makes the return a new SequenceArray object instead. The same idea with
SequenceAlignment.parse_fasta in that it returns a new SequenceAlignment object if and
only if all the sequence lengths are the same. If not, it returns a SequenceArray instead.